Disable reference document link
Reference document link is disabled since document reference from
Access export is too vague to uniquely identify the document that is
subject of the publication. Only agendaitem-treatment can be
identified using this reference.

The regulation type of the publication cannot be used as additional
filter since we don't have document types in Kaleidos DB for documents
originating from Doris.